requirements.txt
streamlit==1.25.0
neo4j==5.9.0
redis==4.5.5
boto3>=1.28.57
botocore>=1.31.57
PyYAML==6.0
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.95.2
pydantic==1.10.11